requirements.txt
annotated-types==0.7.0
audioop-lts==0.2.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.72.0rc1
grpcio-status==1.71.0
gTTS==2.5.4
httplib2==0.22.0
idna==3.10
numpy==2.2.4
opencv-python==4.11.0.86
playsound==1.2.2
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
pygame==2.6.1
pyparsing==3.2.3
requests==2.32.3
rsa==4.9
sounddevice==0.5.1
SpeechRecognition==3.14.2
standard-aifc==3.13.0
standard-chunk==3.13.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
